# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GODNSLOG 2.0 API
  description: |
    GODNSLOG 2.0 API - OAST Interaction Verification & Evidence Platform
    
    This API provides endpoints for:
    - Case management (test tasks, vulnerability verification)
    - Payload generation and tracking
    - Interaction capture (DNS, HTTP, SMTP, LDAP, etc.)
    - Evidence chain generation
    - APIKey management with scopes
    - Audit logging
    
    ## Authentication
    Most endpoints require authentication via:
    - JWT Token (user login)
    - APIKey (programmatic access)
    
    ## Version
    Current version: 2.0.0
    
    ## Base URL
    Production: https://api.godnslog.com/v2
    Development: http://localhost:8080/api/v2
  version: 2.0.0
  contact:
    name: GODNSLOG Team
    email: support@godnslog.com
  license:
    name: MIT
    url: https://github.com/chennqqi/godnslog/blob/master/LICENSE
servers:
  - url: http://localhost:8080/api/v2
    description: Development server
  - url: https://api.godnslog.com/v2
    description: Production server
...
paths:
...
  /interactions:
    get:
      tags:
        - interactions
      summary: List interactions
      description: Retrieve list of interactions with filtering
      operationId: listInteractions
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: case_id
          in: query
          description: Filter by case ID
          schema:
            type: string
        - name: payload_id
          in: query
          description: Filter by payload ID
          schema:
            type: string
        - name: type
          in: query
          description: Filter by interaction type
          schema:
            type: string
            enum: [dns, http, smtp, ldap, smb, ftp]
        - name: start_time
          in: query
          description: Filter by start time (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          description: Filter by end time (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          description: Items per page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of interactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - interactions
      summary: Delete interactions
      description: Delete interactions by IDs
      operationId: deleteInteractions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteRequest'
      responses:
        '204':
          description: Interactions deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token from login
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key for programmatic access
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: Error code
        message:
          type: string
          description: Error message
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp
...
    Interaction:
      type: object
      required:
        - id
        - type
        - timestamp
        - source_ip
      properties:
        id:
          type: string
          description: Interaction ID
        type:
          type: string
          enum: [dns, http, smtp, ldap, smb, ftp]
        case_id:
          type: string
        payload_id:
          type: string
        token:
          type: string
        timestamp:
          type: string
          format: date-time
        source_ip:
          type: string
        # DNS specific
        domain:
          type: string
        dns_type:
          type: string
        # HTTP specific
        method:
          type: string
        path:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
        body:
          type: string
        user_agent:
          type: string
        content_type:
          type: string
        # Common
        raw_data:
          type: string
          description: Raw interaction data
...
    InteractionListResponse:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: OK
        data:
          type: object
          required:
            - items
            - total
            - page
            - page_size
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Interaction'
            total:
              type: integer
            page:
              type: integer
            page_size:
              type: integer
    DeleteRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          items:
            type: string
          description: List of IDs to delete
...